python_app/requirements.txt
Flask==2.3.3
Jinja2==3.1.2
Werkzeug==2.3.7
SQLAlchemy==1.4.49
psycopg2-binary==2.9.9
requests==2.31.0
PyJWT==2.8.0
cryptography==42.0.5
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0